+++ examples/nest-graphql-typeorm/src/sub-task/sub-task.resolver.ts
@@ -1,10 +1,12 @@
+import { QueryService } from '@nestjs-query/core';
 import { CRUDResolver } from '@nestjs-query/query-graphql';
 import { Resolver } from '@nestjs/graphql';
+import { InjectTypeOrmQueryService } from '@nestjs-query/query-typeorm';
 import { TodoItemDTO } from '../todo-item/dto/todo-item.dto';
 import { SubTaskDTO } from './dto/sub-task.dto';
 import { CreateSubTaskDTO } from './dto/subtask-input.dto';
 import { SubTaskUpdateDTO } from './dto/subtask-update.dto';
-import { SubTaskService } from './sub-task.service';
+import { SubTaskEntity } from './sub-task.entity';
 
 @Resolver(() => SubTaskDTO)
 export class SubTaskResolver extends CRUDResolver(SubTaskDTO, {
@@ -14,7 +16,7 @@ export class SubTaskResolver extends CRUDResolver(SubTaskDTO, {
     one: { todoItem: { DTO: TodoItemDTO, disableRemove: true } },
   },
 }) {
-  constructor(readonly service: SubTaskService) {
+  constructor(@InjectTypeOrmQueryService(SubTaskEntity) readonly service: QueryService<SubTaskEntity>) {
     super(service);
   }
 }