+++ examples/nest-graphql-typeorm/src/todo-item/todo-item.service.ts
@@ -1,13 +1,15 @@
-import { QueryService } from '@nestjs-query/core';
-import { TypeOrmQueryService } from '@nestjs-query/query-typeorm';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { AssemblerQueryService, QueryService } from '@nestjs-query/core';
+import { InjectTypeOrmQueryService } from '@nestjs-query/query-typeorm';
 import { TodoItemDTO } from './dto/todo-item.dto';
+import { TodoItemAssembler } from './todo-item.assembler';
 import { TodoItemEntity } from './todo-item.entity';
 
 @QueryService(TodoItemDTO)
-export class TodoItemService extends TypeOrmQueryService<TodoItemDTO, TodoItemEntity> {
-  constructor(@InjectRepository(TodoItemEntity) readonly repo: Repository<TodoItemEntity>) {
-    super(repo);
+export class TodoItemService extends AssemblerQueryService<TodoItemDTO, TodoItemEntity> {
+  constructor(
+    assembler: TodoItemAssembler,
+    @InjectTypeOrmQueryService(TodoItemEntity) queryService: QueryService<TodoItemEntity>,
+  ) {
+    super(assembler, queryService);
   }
 }