+++ packages/core/src/index.ts
@@ -1,6 +1,12 @@
 /* eslint-disable import/export */
 export * from './interfaces';
 export * from './common';
-export { QueryService, AssemblerService, getQueryServiceDTO } from './services';
+export { QueryService, AssemblerQueryService } from './services';
 export { transformFilter, transformQuery, transformSort, QueryFieldMap } from './helpers';
-export { ClassTransformerAssembler, getAssembler, DefaultAssembler, AbstractAssembler, Assembler } from './assemblers';
+export {
+  ClassTransformerAssembler,
+  DefaultAssembler,
+  AbstractAssembler,
+  Assembler,
+  AssemblerFactory,
+} from './assemblers';