+++ packages/query-typeorm/src/decorators/decorators.utils.ts
@@ -1,0 +1,5 @@
+import { Class } from '@nestjs-query/core';
+
+export function getTypeOrmQueryServiceKey<Entity>(entity: Class<Entity>): string {
+  return `${entity.name}TypeOrmQueryService`;
+}