+++ packages/query-typeorm/src/decorators/index.ts
@@ -1,0 +1,2 @@
+export { InjectTypeOrmQueryService } from './inject-typeorm-query-service.decorator';
+export { getTypeOrmQueryServiceKey } from './decorators.utils';