+++ packages/query-typeorm/src/decorators/inject-typeorm-query-service.decorator.ts
@@ -1,0 +1,6 @@
+import { Class } from '@nestjs-query/core';
+import { Inject } from '@nestjs/common';
+import { getTypeOrmQueryServiceKey } from './decorators.utils';
+
+export const InjectTypeOrmQueryService = <Entity>(entity: Class<Entity>): ParameterDecorator =>
+  Inject(getTypeOrmQueryServiceKey(entity));