+++ packages/query-typeorm/src/index.ts
@@ -1,1 +1,3 @@
 export { TypeOrmQueryService } from './services';
+export { InjectTypeOrmQueryService, getTypeOrmQueryServiceKey } from './decorators';
+export { NestjsQueryTypeOrmModule } from './module';